Dimens.xml created and moved the dimen files to there. String format class created to display the value in BRL. Result Fragment now has a better layout. Fixed some errors on previous layout files and removed unused files on the entire project. Changed INT to FLOAT on Model classes, as well on the Arguments passed by the SafeArgs.

+++ app/src/main/java/com/souza/billsapp/data/Expense.kt
@@ -4,7 +4,7 @@ import com.google.firebase.Timestamp
 import com.google.firebase.firestore.ServerTimestamp
 
 data class Expense(
-    val value: Int? = null,
+    val value: Float? = null,
     val description: String? = null,
     @ServerTimestamp
     val date : Timestamp? = null,
